StorylessT63 Fix main screen not showing issue
